Add files via upload
The sentiment is an opinion of the person 
which a person expresses on the particular 
social media site. To get to know the intentions of the person on a particular thing, calculating the sentimental value of the person’s review is must. TSA- Twitter Sentiment Analysis is a tool that will be calculating the subjective and 
polarized value of a particular tweet apart 
from it, it will also fetch the recently posted 
tweets on the particular topic and will also 
calculate their sentimental value which will help to identify on which topic positive or negative tweets are being made alongside it will give a graphical representation of tweets distribution as well. TSA will also tell the sentimental value of the user entering text regardless of that being a real-time tweet on Twitter.